lc0-mcts-engine/requirements.txt
numpy>=1.21.0
python-chess>=1.999
onnxruntime>=1.15.0
tensorflow>=2.13.0; platform_system != "Windows"


